requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.5,<3
pydantic-settings>=2.3,<3
python-dotenv>=1.0,<2
PyJWT>=2.8,<3
cryptography>=42,<43
SQLAlchemy[asyncio]>=2.0,<3
aiosqlite>=0.20,<1
qrcode[pil]>=7.4,<8

# Dev/test
pytest>=8.0,<9
httpx>=0.27,<1

